Moved functional tests from unit test files
