+++ tests/covalent_tests/choose_executor_test.py
@@ -19,10 +19,8 @@
 # Relief from the License may be granted by purchasing a commercial license.
 
 """
-Integration test for choosing executors.
+Test for choosing executors.
 """
-
-import pytest
 
 import covalent as ct
 
@@ -32,42 +30,3 @@ def test_executors_exist():
 
     executor_list = ct.executor._executor_manager.list_executors(print_names=False)
     assert len(executor_list) > 0
-
-
-def test_using_executor_names():
-    """Test that all loaded executors can be used in a simple electron."""
-
-    for executor_name in ["local", "dask"]:
-
-        @ct.electron(executor=executor_name)
-        def passthrough(x):
-            return x
-
-        @ct.lattice()
-        def workflow(y):
-            return passthrough(x=y)
-
-        dispatch_id = ct.dispatch(workflow)(y="input")
-        output = ct.get_result(dispatch_id, wait=True)
-
-        assert output.result == "input"
-
-
-@pytest.mark.skip(reason="Trying to pickle _asyncio.Task even though there is no such object")
-def test_using_executor_classes():
-    """Test creating executor objects and using them in a simple electron."""
-
-    for executor in [ct.executor.LocalExecutor(), ct.executor.DaskExecutor()]:
-
-        @ct.electron(executor=executor)
-        def passthrough(x):
-            return x
-
-        @ct.lattice
-        def workflow(y):
-            return passthrough(x=y)
-
-        dispatch_id = ct.dispatch(workflow)(y="input")
-        output = ct.get_result(dispatch_id, wait=True)
-
-        assert output.result == "input"
