+++ tests/functional_tests/dispatcher_server_test.py
@@ -20,7 +20,6 @@
 
 import covalent as ct
 import covalent._results_manager.results_manager as rm
-from covalent_dispatcher._db.dispatchdb import DispatchDB
 
 
 @ct.electron
